requirements.txt
Django==2.1.12
django-appconf==1.0.2
django-ckeditor==5.4.0
django-compressor==2.2
django-imagekit==4.0.2
django-filter==2.2.0
django-filters-mixin==0.1.4
django-jet==1.0.7
django-js-asset==1.1.0
django-localflavor==2.1
django-polymorphic==2.0.3
django-qrcode==0.3
django-sass-processor==0.7.1
django-tastypie==0.14.2
fcm-django==0.2.13
idna==2.7
libsass==0.15.0
mysqlclient==1.3.12
olefile==0.45.1
pilkit==2.0
Pillow==4.2.1
pyDes==2.0.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.3
rcssmin==1.0.6
requests==2.20.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
six==1.11.0
